major rework based on shrugs suggestion, remove changes made to mintableToken.sol
