+++ contracts/access/ERC20Airdroper.sol
@@ -1,20 +1,28 @@
 pragma solidity ^0.4.24;
 
-import "./MintableToken.sol";
-import "../../ECRecovery.sol";
+import "../ECRecovery.sol";
+import "../math/SafeMath.sol";
+import "../token/ERC20/MintableToken.sol";
 
 /**
- * @title AirdropableToken
- * @author SylTi
- * @dev Simple ERC20 Token that allow for an airdrop to be realised based on signatures from the owner.
+ * @title ERC20Airdroper
+ * @author SylTi & Shrugs
+ * @dev Simple contract that allow for an airdrop to be realised based on signatures from the owner and a MintableToken
  * This allow the owners of the token to forward the gas cost to the receiver of the airdrop instead of having to pay for it.
  */
 
 
-contract AirdropableToken is MintableToken {
+contract ERC20Airdroper is Ownable {
   using ECRecovery for bytes32;
+  using SafeMath for uint256;
 
+  MintableToken token;
   mapping (address => uint256) public nonces;
+
+
+  constructor(MintableToken _token) public {
+    token = _token;
+  }
 
   /**
    * @dev Function to mint tokens given a valid signature.
@@ -24,13 +32,12 @@ contract AirdropableToken is MintableToken {
    * @param _signature Signature of the hash
    * @return A boolean that indicates if the operation was successful.
    */
-  function claim(
+  function mint(
     address _to,
     uint256 _amount,
     uint256 _nonce,
     bytes _signature
   )
-    canMint
     public
     returns (bool)
   {
@@ -39,7 +46,8 @@ contract AirdropableToken is MintableToken {
     address signer = proof.recover(_signature);
     require(signer == owner, "Not signed by owner");
     nonces[_to] = _nonce.add(1);
-    return super.doMint(_to, _amount);
+    require(token.mint(_to, _amount), "minting failed");
+    return true;
   }
 
   /**
@@ -63,5 +71,14 @@ contract AirdropableToken is MintableToken {
         _nonce
       )
     ).toEthSignedMessageHash();
+  }
+
+  /**
+  * @dev Function to stop minting new tokens.
+  * @return True if the operation was successful.
+  */
+  function finishMinting() onlyOwner public returns (bool) {
+    require(token.finishMinting(), "finish minting failed");
+    return true;
   }
 }
