+++ contracts/token/ERC20/MintableToken.sol
@@ -10,12 +10,12 @@ import "../../ownership/Ownable.sol";
  * Based on code by TokenMarketNet: https://github.com/TokenMarketNet/ico/blob/master/contracts/MintableToken.sol
  */
 contract MintableToken is StandardToken, Ownable {
-
   event Mint(address indexed to, uint256 amount);
   event MintFinished();
 
   bool public mintingFinished = false;
-
+  
+  
   modifier canMint() {
     require(!mintingFinished);
     _;
@@ -27,7 +27,7 @@ contract MintableToken is StandardToken, Ownable {
   }
 
   /**
-   * @dev Function to mint tokens from approved address
+   * @dev Function to mint tokens
    * @param _to The address that will receive the minted tokens.
    * @param _amount The amount of tokens to mint.
    * @return A boolean that indicates if the operation was successful.
@@ -41,7 +41,11 @@ contract MintableToken is StandardToken, Ownable {
     public
     returns (bool)
   {
-    return doMint(_to, _amount);
+    totalSupply_ = totalSupply_.add(_amount);
+    balances[_to] = balances[_to].add(_amount);
+    emit Mint(_to, _amount);
+    emit Transfer(address(0), _to, _amount);
+    return true;
   }
   
   /**
@@ -53,25 +57,4 @@ contract MintableToken is StandardToken, Ownable {
     emit MintFinished();
     return true;
   }
-
-  /**
-   * @dev Private function that do the actual minting of tokens
-   * @param _to The address that will receive the minted tokens.
-   * @param _amount The amount of tokens to mint.
-   * @return A boolean that indicates if the operation was successful.
-   */
-  function doMint(
-    address _to,
-    uint256 _amount
-  )
-    internal
-    returns (bool)
-  {
-    totalSupply_ = totalSupply_.add(_amount);
-    balances[_to] = balances[_to].add(_amount);
-    emit Mint(_to, _amount);
-    emit Transfer(address(0), _to, _amount);
-    return true;
-  }
-
 }
